+++ libs/select-language/README.md
@@ -1,0 +1,7 @@
+# select-language
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test select-language` to execute the unit tests via [Jest](https://jestjs.io).
