+++ libs/select-language/jest.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+  displayName: "select-language",
+  preset: "../../jest.preset.js",
+  transform: {
+    "^.+\\.[tj]sx?$": [
+      "babel-jest",
+      { cwd: __dirname, configFile: "./babel-jest.config.json" },
+    ],
+  },
+  moduleFileExtensions: ["ts", "tsx", "js", "jsx"],
+  coverageDirectory: "../../coverage/libs/select-language",
+};
