+++ libs/select-language/src/lib/select-language.spec.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { render } from "@testing-library/react";
+
+import SelectLanguage from "./select-language";
+
+describe("SelectLanguage", () => {
+  it("should render successfully", () => {
+    const { baseElement } = render(<SelectLanguage />);
+    expect(baseElement).toBeTruthy();
+  });
+});
