+++ workspace.json
@@ -193,6 +193,31 @@
           }
         }
       }
+    },
+    "select-language": {
+      "root": "libs/select-language",
+      "sourceRoot": "libs/select-language/src",
+      "projectType": "library",
+      "targets": {
+        "lint": {
+          "executor": "@nrwl/linter:eslint",
+          "options": {
+            "lintFilePatterns": [
+              "libs/select-language/**/*.{ts,tsx,js,jsx}"
+            ]
+          }
+        },
+        "test": {
+          "executor": "@nrwl/jest:jest",
+          "outputs": [
+            "coverage/libs/select-language"
+          ],
+          "options": {
+            "jestConfig": "libs/select-language/jest.config.js",
+            "passWithNoTests": true
+          }
+        }
+      }
     }
   },
   "cli": {
